* 'develop' of bcn-mb-git:dragon/client: (23 commits)
  Fixes
  force blood when latching and grabbing to start red
  Village air scene fix
  Fixes
  FEATURE SETTINGS: "particles" parameter added to all profiles.
  Update water cave scene
  Rays texture
  gear placeholder fix
  Fixes
  mines
  Gear proto
  Cleanup, fixes and low version of monkey cloud effect
  check event
  Custom Particle System prewarm option
  Global Event: Score is integer, changed current value to globalScore. WIP
  Tid - Add new lower_damage_origin tid
  updated knockback value
  mines wagon
  Menu scene, wip low effect
  Polish
  ...


Former-commit-id: 873680a039edbb2fb1e22c8bf88f656a0b3bbe21

+++ Assets/Game/Core/GameServerManagerCalety.cs
@@ -365,7 +365,7 @@ public class GameServerManagerCalety : GameServerManager {
 	/// <param name="_eventID">The identifier of the target event.</param>
 	/// <param name="_score">The score to be registered.</param>
 	/// <param name="_callback">Callback action.</param>
-	override public void GlobalEvent_RegisterScore(int _eventID, float _score, ServerCallback _callback) {
+	override public void GlobalEvent_RegisterScore(int _eventID, int _score, ServerCallback _callback) {
 		// Compose parameters and enqeue command
 		Dictionary<string, string> parameters = new Dictionary<string, string>();
 		parameters.Add("eventId", _eventID.ToString(JSON_FORMAT));
